Minor update to 'letting users report a technical problem' guidance
Small change agreed with Dave Barrett and Robin Barker, as part of a conversation about error handling. It's less a case of designers checking whether service management are already tracking an error, and more about having a conversation with them to see if directing users to the contact form is helpful in a given scenario.